Algunas correcciones en clases genericas, scripts y menu

+++ P2_TPI/src/p2_tpi/Dao/GenericDAO.java
@@ -9,7 +9,7 @@ public interface GenericDAO<T> {
     void insertTx(T entidad, Connection conn) throws Exception;
     void actualizar(T entidad)throws Exception;
     void eliminar(long id)throws Exception;
-    T getById(int id)throws Exception;
+    T getById(long id)throws Exception;
     List<T> getAll()throws Exception;
 
 }
